+++ packages/terra-application-layout/src/terra-dev-site/doc/common/UtilityOption.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import AppDelegate from 'terra-app-delegate';
 import ContentContainer from 'terra-content-container';
-import ActionHeader from 'terra-clinical-action-header';
+import ActionHeader from 'terra-action-header';
 
 const propTypes = {
   name: PropTypes.string,