+++ packages/terra-application-layout/CHANGELOG.md
@@ -13,6 +13,9 @@ Unreleased
 ### Removed
 Removed dependency on terra-clinical-action-header
 
+### Changed
+* Add .module extension to SCSS file(s)
+
 2.8.0 - (July 10, 2018)
 ------------------
 ### Changed