+++ packages/terra-application-layout/CHANGELOG.md
@@ -5,6 +5,13 @@ Unreleased
 ----------
 ### Added
 * Added `sv` and `sv-SE` translations
+* Added dependency on action-header
+
+### Changed
+* Replaced clinical-action-header with action-header
+
+### Removed
+Removed dependency on terra-clinical-action-header
 
 ### Changed
 * Add .module extension to SCSS file(s)