+++ packages/terra-application-layout/CHANGELOG.md
@@ -4,8 +4,25 @@ Changelog
 Unreleased
 ----------
 ### Added
-* Added `sv` and `sv-SE` translations
 * Expanded support for navigational paths.
+
+2.10.0 - (July 26, 2018)
+------------------
+### Changed
+* Minor dependency updates
+
+2.9.0 - (July 25, 2018)
+------------------
+### Added
+* Added `sv` and `sv-SE` translations
+* Added dependency on action-header
+
+### Changed
+* Replaced clinical-action-header with action-header
+* Add .module extension to SCSS file(s)
+
+### Removed
+Removed dependency on terra-clinical-action-header
 
 2.8.0 - (July 10, 2018)
 ------------------